SNOW-1284606 Move running_on_* to tests/utils.py
