+++ tests/integ/modin/series/test_rename.py
@@ -14,9 +14,9 @@ import pytest
 from modin.pandas import Index, MultiIndex, Series
 
 import snowflake.snowpark.modin.plugin  # noqa: F401
-from tests.integ.conftest import running_on_public_ci
 from tests.integ.modin.sql_counter import SqlCounter, sql_count_checker
 from tests.integ.modin.utils import assert_index_equal, assert_series_equal
+from tests.utils import running_on_public_ci
 
 
 class TestRename:
